requirements.txt
pandas
scikit-learn
scipy
apache-airflow
mlflow
requests
pyyaml
spotipy
pandas
numpy
fastapi
uvicorn